requirements/dev.txt
-r common.txt
git+git://github.com/kaleidos/django-sampledatahelper.git
django-debug-toolbar
